requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
opencv-python==4.8.0.76
face-recognition==1.3.0
numpy==1.24.3
Werkzeug==2.3.7
flask-socketio==5.3.6
gunicorn==20.1.0
eventlet==0.33.3 